+++ CHANGELOG.md
@@ -37,6 +37,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
    * Allow `Instrument.rename` to take a fuction or mapping dict as input,
      after adapting routine to use `Meta.rename`
    * Added an update method and type evaluation method to MetaLabels.
+   * Added a new MetaHeader class for holding global metadata.
+   * Added a pysat Instrument for loading pysat-written NetCDF4 files.
    * Allowed MetaLabels to be expanded through setting new Meta data values.
 * Deprecations
    * Removed `freq` as a standard kwarg for `pysat.Instruments.download`
@@ -113,6 +115,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
    * Organized and reduced duplication in the Meta test class
    * Added CI reports for supported data products
    * Added a cap on coveralls to ensure success of continuous integration
+   * Updated tests in `test_meta` to search all warnings, not just the first
 
 [3.0.1] - 2021-07-28
 --------------------